+++ src/main/java/ru/otus/danilchenko/algorithms/lesson5/ChessPiece.java
@@ -2,6 +2,6 @@ package ru.otus.danilchenko.algorithms.lesson5;
 
 public interface ChessPiece {
     int getNumberOfSteps();
-    long getPosition();
+    int getPosition();
     long getStepsPosition();
 }
